MSI-783: Add core service GetProductTypeBySku - move service to InventoryCatalog
