+++ app/code/Magento/InventoryCatalog/Api/GetProductTypeBySkuInterface.php
@@ -5,9 +5,7 @@
  */
 declare(strict_types=1);
 
-namespace Magento\Catalog\Api;
-
-use Magento\Framework\Exception\InputException;
+namespace Magento\InventoryCatalog\Api;
 
 /**
  * Get product types id by product skus.
@@ -20,7 +18,6 @@ interface GetProductTypeBySkuInterface
      * Returns product types id by product skus.
      *
      * @param array $skus
-     * @throws InputException
      * @return array (key: 'sku', value: 'product_type')
      */
     public function execute(array $skus);
