+++ app/code/Magento/InventoryCatalog/Test/_files/products_all_types_rollback.php
